backend/requirements.txt
flask
flask-cors
openai
pandas
requests
python-dotenv
